[autoscaler] add timezone in policy edit page
