+++ package.json
@@ -61,12 +61,13 @@
     "@ngrx/store": "^7.2.0",
     "@ngrx/store-devtools": "^7.2.0",
     "@swimlane/ngx-charts": "^10.1.0",
+    "@types/moment-timezone": "^0.5.12",
     "angular2-virtual-scroll": "^0.3.1",
     "core-js": "^2.6.5",
     "hammerjs": "^2.0.8",
     "intersect": "^1.0.1",
     "moment": "^2.24.0",
-    "moment-timezone": "^0.5.23",
+    "moment-timezone": "^0.5.12",
     "ngx-moment": "^3.3.0",
     "normalizr": "^3.2.3",
     "reselect": "^3.0.1",
