+++ src/frontend/packages/core/src/features/applications/application/application-tabs-base/tabs/autoscaler-tab/autoscaler-tab.component.html
@@ -116,7 +116,7 @@
         <app-tile>
           <mat-card>
             <mat-card-header>
-              <mat-card-title>Scheduled Limit Rules</mat-card-title>
+              <mat-card-title>Scheduled Limit Rules in {{policy.schedules.timezone}}</mat-card-title>
             </mat-card-header>
             <mat-card-content>
               <div class="app-metadata">
