+++ src/frontend/packages/core/src/shared/components/list/list-types/app-autoscaler-metric-chart/app-autoscaler-metric-chart-card/app-autoscaler-metric-chart-card.component.ts
@@ -17,7 +17,7 @@ import { PaginationMonitorFactory } from '../../../../../../shared/monitors/pagi
 import {
   AppAutoscalerAppMetric,
 } from '../../../../../../../../store/src/types/app-autoscaler.types';
-import { buildLegendData } from '../../../../../../../../store/src/helpers/autoscaler-helpers';
+import { buildLegendData } from '../../../../../../../../store/src/helpers/autoscaler/autoscaler-util';
 
 @Component({
   selector: 'app-app-autoscaler-metric-chart-card',
